🚚 Move poll leaf components into features/poll (RAL-1291)
- Move invite-dialog, vote-summary-progress-bar and create-poll from
  components/ to features/poll/components/
- Delete poll-status.tsx: PollStatusBadge has no importers (dead code)
- Remove the freed entries from the biome.json migration allowlist
