+++ apps/web/biome.json
@@ -124,11 +124,9 @@
         "src/components/discussion/discussion.tsx",
         "src/components/event-card.tsx",
         "src/components/forms/poll-settings.tsx",
-        "src/components/invite-dialog.tsx",
         "src/components/layouts/poll-layout.tsx",
         "src/components/logo.tsx",
         "src/components/new-participant-modal.tsx",
-        "src/components/poll-status.tsx",
         "src/components/poll/desktop-poll.tsx",
         "src/components/poll/desktop-poll/participant-row.tsx",
         "src/components/poll/manage-poll.tsx",
