+++ apps/web/src/features/poll/components/vote-summary-progress-bar.tsx
